Two things: - Split the unit tests into their own test file, making it easier to see which tests are "live" and which are just in-memory - Added in more tests around recovery scenarios, especially since it's changing quite a bit.
The big changes are that we can potentially upgrade a link recovery to a connection recovery, depending on what has happened when trying to recover the link. This will work properly even if the user cancels - if that happens we'll just close everything and it'll all come back the next time the user calls in.
